+++ src/adapters/sqlite/encodeSchema/test.js
@@ -9,7 +9,7 @@ const testSchema = appSchema({
       name: 'tasks',
       columns: [
         { name: 'author_id', type: 'string', isIndexed: true },
-        { name: 'position', type: 'number', isOptional: true, isIndexed: true },
+        { name: 'order', type: 'number', isOptional: true, isIndexed: true },
         { name: 'created_at', type: 'number' },
       ],
     }),
@@ -21,12 +21,12 @@ const testSchema = appSchema({
 })
 
 const expectedSchema =
-  'create table tasks (_changed, _status, id primary key, last_modified, author_id, position, created_at);' +
-  'create index tasks_author_id on tasks (author_id);' +
-  'create index tasks_position on tasks (position);' +
-  'create index tasks__status on tasks (_status);' +
-  'create table comments (_changed, _status, id primary key, last_modified, is_ended, reactions);' +
-  'create index comments__status on comments (_status);'
+  'create table "tasks" ("_changed", "_status", "id" primary key, "last_modified", "author_id", "order", "created_at");' +
+  'create index tasks_author_id on "tasks" ("author_id");' +
+  'create index tasks_order on "tasks" ("order");' +
+  'create index tasks__status on "tasks" ("_status");' +
+  'create table "comments" ("_changed", "_status", "id" primary key, "last_modified", "is_ended", "reactions");' +
+  'create index comments__status on "comments" ("_status");'
 
 describe('watermelondb/adapters/sqlite/encodeSchema', () => {
   it('encodes schema', () => {